Changed the mana system and updated spell costs
In the previous iteration, mana costs were randomly assigned on map
generation to be a value between 0 and 9. Additionally, spells took multiple
units of mana to cast.

However, this was problematic for two reasons. First, the UI became difficult to read, as a number in the background of each tile made the map look rather busy and it was not possible to read the value when another creature
or object was on the tile. Second, mana was fairly abundant, making it easy
to spam spells such as teleport without caring too much about positioning.

By reducing mana amounts, this cleans up the map UI, increases the
impact of mana manipulation enemies, and hopefully makes the choice of
spell usage and positioning more impactful.